+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 ## Added
+- frontend: Display number of GPUs per node in resources page and node details
+  page.
 - agent: Add total quantity of memory and number of gpus in cluster stats
   response.
 - conf: Select `gres` and `gres_used` attributes by default in `slurmrestd`